Factor out the block mocking.
